creating prescription service
