+++ ClientApp/src/models/medication.ts
@@ -6,3 +6,11 @@ export interface Medication {
   capacity: number;
   price: number;
 }
+
+export interface MedicationDosages {
+  medicationId: string;
+  startDate: string;
+  endDate: string;
+  quantity: number;
+  frequency: number;
+}
